+++ FaridKhaidari/1a_mul/mul_funcs/intr.hh
@@ -1,0 +1,4 @@
+#ifndef INTR_HH
+#define INTR_HH
+
+#endif // INTR_HH
