+++ FaridKhaidari/1a_mul/mul_funcs/opcl.hh
@@ -1,0 +1,4 @@
+#ifndef OPCL_HH
+#define OPCL_HH
+
+#endif // OPCL_HH
