+++ FaridKhaidari/1a_mul/tst/mul1.dat
