+++ FaridKhaidari/1a_mul/tst/mul3.dat
