+++ FaridKhaidari/1a_mul/tst/mul4.dat
