+++ FaridKhaidari/1a_mul/tst/mul5.dat
